requirements.txt
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
Pillow==10.4.0
PyYAML==6.0.2
scikit-image==0.24.0
pandas==2.2.3
matplotlib==3.9.2
tqdm==4.66.5
